requirements.txt
django-s3direct==0.3.9
django-compressor==1.6
django-sass-processor==0.3
django-bower==5.0.4
dj-database-url==0.3.0
dj-static==0.0.6
whitenoise==2.0.6
gunicorn==19.1.1
psycopg2==2.5.1
django-autoslug==1.9.3
django-vote==1.1.2
python-social-auth==0.2.14
